(maint) DRY up duplicate checking w/ events/venues
Prior to this commit, the duplicate checking logic for venues and events
was duplicated between each of the controllers. This reduces
maintainability if ever the duplicate checking algorithm needs to
change. This commit moves the identical logic into the mixin (used only
by those two controllers), abstracting away instance variable names and
models so that they can be adequately determined at runtime.